Move get_langchain_context to prompt.py and align types
